+++ vertexai/preview/language_models.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# Copyright 2022 Google LLC
+# Copyright 2023 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,4 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-__version__ = "1.25.0"  # {x-release-please-version}
+"""Classes for working with language models."""
+
+from vertexai.language_models._language_models import (
+    ChatModel,
+    ChatSession,
+    InputOutputTextPair,
+    TextEmbedding,
+    TextEmbeddingModel,
+    TextGenerationModel,
+    TextGenerationResponse,
+)
+
+__all__ = [
+    "ChatModel",
+    "ChatSession",
+    "InputOutputTextPair",
+    "TextEmbedding",
+    "TextEmbeddingModel",
+    "TextGenerationModel",
+    "TextGenerationResponse",
+]